requirements.txt
# requirements.txt
streamlit==1.44.1
pandas==2.2.3
numpy==2.0.2
openai==1.76.0
python-dotenv==1.1.0
plotly==6.0.1
pyarrow==19.0.1
jinja2==3.1.6
typing-extensions==4.13.2